Fixed overflow issue that would only happen on 64 bit compiles with hash tables or geobuckets larger than 2^31. After this change, all remaining issues found by the static analysis of Visual Studio 2012 Professional are false positives.

+++ src/mathicgb/ChainedHashTable.hpp
@@ -98,7 +98,7 @@ private:
 template <typename HashControl>
 ChainedHashTable<HashControl>::ChainedHashTable(HashControl M, int nbits)
   : M_(M),
-    table_size_(1 << nbits)
+    table_size_(static_cast<size_t>(1) << nbits)
 {
   hashtable_.resize(table_size_);
